Generalized and refactored the locking in Rados backend.
The locking is now handled in a single function.
This functions is also used in the asynchronous operations. This operations have
been simplified in this operation: we now do a synchronous stat after the locking.
Previously stat was the first asynchronous operation. This will slightly slow down
the async operations (for the gain of code simplicity).