Moved Rados locking to aio_unwatch.
The unwatch step is pretty slow, so the notification structure is now in a seprate
internal object, which is left to be deleted by the callback of aio_unwatch. We need
to keep the structure around for that time as notifications could still arrive
between the call to aio_unwatch and the actual unwatching.