[ssi_af_ls] Creates buffer object with both ostream and char* interfaces
XrdSsi requires that we return data to it as a char* with a size, but
this is not a convenient object for us to fill with data. By creating a
streambuf/ostream wrapper around a char[] buffer we can give XrdSsi what
it needs and use the more convenient stream operators in our own code.

The classes used in this buffer implementation should be moved into the
generic code when I have got the whole thing working.