Disabled local garbage collection of unprocessed jobs.
The jobs will now remain attached to the agent of the drive process. The garbage collector
process will now requeue them instead of drive process. This will allow the drive process
to restart faster for a new session and the drive process to requeue the jobs faster a it
processes jobs in bulk.