Optimized retrive request queueing. Added logs.
The retrieve request being queued in one queue only, we can add a reference to
it in the queue, and then create the request object after. This is a simplification
from the previous algorithm.

Added logs for request creation. This will now create a log line where both the object
name and the file Id and path are available as parameters, easing debugging,.