Split queuing archive requests in top and bottom half.
The top half goes as far as when the request is safe in the object store.
At that point, the bottom half is launched in a new thread, and success is returned
the called. This will enable lower latency for the users while retaining
the same data safety.

This version is experimental aas it spans an undetermined number of threads. A more controlled
version with a work queue should be implenented on the long run.