[lpa-stream] Exposes OStoreDB::getArchiveQueueItor() to the application
I tried putting this through the Scheduler, but as getArchiveQueueItor
returns an object of type QueueItor, I would have to expose the itor
class everywhere that the Scheduler is used. As this would not be
easy to maintain, I backtracked and instead call the method directly
on the OStoreDB object.