[lpa-stream] Adds move constructor
Default move constructor did not work in case of an empty list, as in
this case the list is not moved, a new empty list is instantiated. So
the copied iterators are invalid.

New move constructor is the same as default with an additional check to
reset the iterators if the lists to be moved are empty.