Upgraded to EOS 4.2.23-20180607211817gitc44922b
This version of EOS has two fixes:

This version of EOS fixes the problem where stagerrm was removing
disk copies from the EOS namespace but was not removing the
physical files from the FSTs.

This version of EOS fixes the problem where the crashing of a tape
server whilst was creating a disk replica would cause the FST to
ask the MGM to completely remove the EOS file from the EOS
namespace, in other words dataloss.