Reduced the number of hardcoded reties for a given file.
This is a stop gap solution for FTS tests which generate too many mounts as many file transfers
fail on checksum check and the file gets delete after the archive requests get generated.

On the longer run, we will need to imlement a placeholder mechanism where the taper server
can skip missing files and carry on with the migration session without unmounting the tape.