Implemented promotion of repack requests from Pending to ToExpand
This promotion is controlled so that only a limited number a requests
are in the
state ToExpand or Starting at any point in time. This ensures both the
availabality
of repack file requests to system while preventing an explosion of file
level requests.

Created a one-round popping from the container (algorithms) with status
switching.
  - Used for repack requests switching from pendig to to expand

Added ElementStatus to algorithms.

Implemented promotion interface in Scheduler and OstoreDb. The actual
decision is taken at
the Scheduler level. The function itself is called by the
RepackRequestManager.

Promotion is tested in a unit test.

Various code maintenance:
Switched to "using"-based constructor inheritance.
Fixed privacy of function in cta::range.