Fixed weakness in BackendRados implmeetation.
Attempting to lock a non-existing object leads to the creation of a zero-sized one. When this happens, the
locking code deletes the freshly created object immediately. Nevertheless, this prevented the creation of a
new object with this name. The creation code now retries to overcome this race.
In addition, reads finding a zero-sized object also consider it as absent.