#533 Changed strategy for implementation.
Moved the space reservation information to the DriveStatus object store object instead of a new central registry.
The central registry would have been a single point of contention as was the DriveRegistry before being split into
DriveStates. As the problem is so close to the one of the drive status, we can actually reuse the drive status for this purpose.

The algorithm will also change as we move the responsibility of querying the free space from the disk systems into the OStoreDb
object instead of the Scheduler. This leads to a slightly worth layering of responsibilities, making the OStoreDb::RetrieveMount
object a client of the disk::DiskSystemFreeSpaceList object.

The current implementation will also query the free space from the disk systems on each pop, instead of doing so in a globally
cached fashion. With the new model, we could cache the free space per drive (if needed), but not globally. This is not expected
to be a real issue and free space is a global counter in the disk system, expected to be readily available.