[migration] Enforces absolute pathnames for EOS dir/file injection
Now we are also tolerant of whether paths in config file do or do
not have a trailing slash.

Pathnames from the DB must have a leading slash if we have to strip off
the CASTOR prefix, but in any case we will convert all filenames to
absolute filenames by adding the EOS prefix.