Revert "[catalogue] cta/CTA#951 Fixed"
This reverts commit 9ba6951dcc674917c909674ce697efe2b2460533.

The orginal commit is being reverted because it intentionally
forced the "list contents of a tape" query to use indicies by
replacing table joins with sub-queries.  These sub-queuries
unfortunately turned out to be slower than the original query
when it was executed in a "clean" environment and the database
live memory cache was cold.  By "clean" environment I mean one
without PARALLEL enabled on the following two tables:

ARCHIVE_FILE
TAPE_FILE

PARALLEL is enabled on these two tables when a CASTOR to CTA
migration takes place.  It was left in place with the hope that
it would not interferre with the day to day running of CTA.  This
has now been proven to not be the case.