disk: don't force xroot client to send stat over the wire
The DiskReadTask task issues DiskFile::size to validate the file size.
As Xroot protocol always return the StatInfo on open and as tape files
are immutable (even if not files are not locked, this calling stat and
then read doesn't guarantee that file is not changed), there are no
needs to enforce the xroot client to ignore stat cache.