Resolve "Remove cta-taped dependency on eos-client"
This commit introduces a new RPM `cta-eosdf` that installs an
external script to query the disk system for free disk space. The
dependency on eos-client is removed from taped, and instead it is the
external script that has a dependency on the eos-client, which it uses
to query for free disk space.
The path to the script is expected to be provided in the taped
configuration.
If there is an error getting the free disk space from the script, then
no backpressure is applied and the queues are not put to sleep; we
still allow staging files for retrieve.