Resolve "Remove cta-taped dependency on eos-client"
This commit introduces a new RPM cta-eos-freediskspace that installs an
external script to query the disk system for free disk space. The
dependency on eos-client is removed from taped, and instead it is the
external script that has a dependency on the eos-client, which it uses
to query for free disk space.
The path to the script is expected to be provided in the taped
configuration.
If there are disk systems but no external script is provided in the
configuration, then an error is logged and the queue is put to sleep.