Fail pipeline immediately if any job fails
This commit adds functionality to fail the running pipeline as soon as any of the jobs fail. This frees up resources and reduces runner contention. For the main branch, this feature is not enabled as there is it often desirable to see all failed jobs.

Closes #852

See merge request cta/CTA!627

Changelog: feature
Issue: 852