Ensure cta-versionlock script can update versionlock.list
The cta-versionlock script was previously renamed to cta-checkpkg with reduced functionality. However, this refactor removed too much functionality and was not able to integrate the `versionlock.cta` file into `verionlock.list`. This went unnoticed as we did not have any package upgrades for a long time. This commit fixes this and adds back the functionality.

Closes #771

See merge request cta/CTA!626

Changelog: maintenance
Issue: #771