Fix liquibase-test failing due to CTA not supporting previous catalogue version
This commit adds a check to the `k8s-liquibase-test` to see if there are multiple supported catalogue versions in CTA. If there is only 1 supported version, there is no need to run the test (and the test would always fail). As such, this commit stops the test from running if there is only 1 supported test.

Closes #966

See merge request cta/CTA!703

Changelog: bug
Issue: #966