Further clean-up and performance improvements to the Helm setup
This commit removes the init chart by moving some definitions around. This means we are left with the kdc chart in the beginning, which can be installed at the same time as the catalogue and scheduler charts. Further cleanup of the init scripts, the spec/ directory in docker/ctafrontend/, and some tiny bug fixes are also included.

Closes #954

See merge request cta/CTA!698

Changelog: maintenance
Issue: #954