[frontend] Set value of grpc server processing threads from config
Previously we were ignoring the value set in the configuration file
for the number of server processing threads and always setting the
default value instead. This commit ensures that if a proper value is
set in the config, then it is used instead of the default. It also adds
a validity check for the value set, and throws an exception at startup
time if the check does not pass.